Move composer Dropdown from features/composer to features/compose
